Move homepage specific components to homepage folder, and only keep general components/containers in root eponym folders
