+++ src/homepage/mainContent/explanations/HomePageExplanations.jsx
