+++ src/homepage/mainContent/explanations/HomePageExplanationsBlock.jsx
