+++ src/homepage/mainContent/explanations/discussTeam.png
