+++ src/homepage/mainContent/title/HomePageTitle.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import TitleH1 from '../components/TitleH1';
+import TitleH1 from '../../../components/TitleH1';
 
 class HomePageTitle extends React.Component {
   render() {
